fix: set minimum oauth linked account

+++ src/main/java/org/swmaestro/repl/gifthub/auth/service/UserService.java
@@ -164,6 +164,9 @@ public class UserService implements UserDetailsService {
 	}
 
 	public OAuth deleteOAuthInfo(User user, OAuthPlatform oAuthPlatform) {
+		if (oAuthService.count(user) <= 1) {
+			throw new BusinessException("최소 하나 이상의 OAuth 연동 계정이 존재해야 합니다.", StatusEnum.BAD_REQUEST);
+		}
 		return oAuthService.delete(user, oAuthPlatform);
 	}
